+++ lib/src/parser/literals.rs
@@ -20,20 +20,3 @@ pub use number::{number, Number};
 pub use string::{quoted_single_line_string, string};
 #[doc(inline)]
 pub use whitespace::{inline_whitespace, newline, whitespace};
-
-use nom::types::CompleteStr;
-use nom::{do_parse, named, tag};
-
-named!(
-    pub null(CompleteStr) -> (),
-    do_parse!(
-        tag!("null")
-        >> (())
-    )
-);
-
-#[test]
-fn parses_for_null() {
-    use crate::utils::test::ResultUtilsString;
-    null(CompleteStr("null")).unwrap_output();
-}
