+++ lib/src/serde/de.rs
@@ -12,6 +12,7 @@ use serde::forward_to_deserialize_any;
 use serde::Deserialize;
 
 use crate::parser;
+use crate::parser::null::null;
 use crate::parser::literals;
 use crate::value;
 
@@ -226,7 +227,7 @@ impl<'de> Deserializer<'de> {
     }
 
     fn parse_null(&mut self) -> Result<(), Error> {
-        let (remaining, ()) = literals::null(self.input)?;
+        let (remaining, ()) = null(self.input)?;
         self.input = remaining;
         Ok(())
     }
